Includes streaming optimized insts mechanism

+++ src/arch/x86/insts/microfpop.hh
@@ -79,7 +79,7 @@ namespace X86ISA
         std::string generateDisassembly(Addr pc,
             const SymbolTable *symtab) const;
 
-        virtual uint8_t getImmediate() { panic("getImmediate() should only be called by RegOpImm, called by FpOp"); }
+        virtual uint64_t getImmediate() { panic("getImmediate() should only be called by RegOpImm, called by FpOp"); }
 
     };
 }
